Setup doc updates (#3)
* OSX: Optimistically drop easy_install, pip, pyobjc; add nvm

Looks like nothing is using it. We can add it back if something in JS is somehow using it, but more likely we should work to drop it as a dependency.

* Drop gitter

Slack > gitter, but that's in the communication doc for now.

* OSX: Optimistically drop node-gyp global dependency

It should be documented in each individual repo if it is actually needed, and preferably not as a global. Further it's not part of Windows/Linux setups at this time.

* Discourage brew installation of nvm as discovered by @kangas

* Similarly to nvm, http://brew.sh/ has different instructions

So we should just point developers in the right direction.

i.e.
/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
vs
ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"