Update setup.md: Add RedHat6 install instructions
Installs everything for me except:
https://github.com/atom/node-keytar

Not exhaustive though as I've only done this once on RHEL 6.7 to try to support Compass on RHEL 6.x, and it's probably a lot easier on RHEL 7.3 or later, but better to get the broad strokes out there and incrementally refine if I need to do this again. The rest which I may have missed should be the typical "LMGTFY" type of error, like the "return binding.open(" one for electron-prebuilt.

Still getting things like: 
https://www.centos.org/forums/viewtopic.php?t=5807

e.g.
[mongouser@localhost compass]$ npm start

> mongodb-compass@1.6.0-dev start /home/mongouser/Projects/compass
> hadron-build develop

/home/mongouser/Projects/compass/node_modules/electron-prebuilt/dist/electron: /usr/lib64/libstdc++.so.6: version `GLIBCXX_3.4.14' not found (required by /home/mongouser/Projects/compass/node_modules/electron-prebuilt/dist/electron)
/home/mongouser/Projects/compass/node_modules/electron-prebuilt/dist/electron: /usr/lib64/libstdc++.so.6: version `CXXABI_1.3.5' not found (required by /home/mongouser/Projects/compass/node_modules/electron-prebuilt/dist/electron)
/home/mongouser/Projects/compass/node_modules/electron-prebuilt/dist/electron: /usr/lib64/libstdc++.so.6: version `GLIBCXX_3.4.15' not found (required by /home/mongouser/Projects/compass/node_modules/electron-prebuilt/dist/electron)
/home/mongouser/Projects/compass/node_modules/electron-prebuilt/dist/electron: /usr/lib64/libstdc++.so.6: version `GLIBCXX_3.4.15' not found (required by /home/mongouser/Projects/compass/node_modules/electron-prebuilt/dist/libnode.so)

But it's a start to work the problem and see if it's resolvable.